requirements.txt
accelerate==1.5.2
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
bitsandbytes==0.45.3
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
filelock==3.18.0
fsspec==2025.3.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
kombu==5.5.0
lxml==5.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2-binary==2.9.10
pypdf==5.4.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==67.6.0
six==1.17.0
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.6.0
tika==2.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
watchdog==2.3.0
wcwidth==0.2.13
gunicorn==20.1.0
beautifulsoup4
langchain